test: add tests for createManyProducts handler
